Renaming MyBox to ReciteBox
Used the following script

```
git grep -l 'MyBox' | xargs gsed -i 's@MyBox@ReciteBox@g'

for file in $(git ls-files | grep 'MyBox')
do
    mkdir -p $(dirname ${file/MyBox/ReciteBox})
    mv ${file} ${file/MyBox/ReciteBox}
done
```
