--- src/components/MyBox/MyBox.css
@@ -1,2 +1,0 @@
-.MyBox {
-}
