--- src/components/MyBox/index.js
@@ -1,3 +1,0 @@
-import MyBox from './MyBox';
-
-export default MyBox;
