+++ src/components/ReciteBox/PhraseNavigator.js
