+++ src/components/ReciteBox/ReciteBox.css
@@ -1,0 +1,2 @@
+.ReciteBox {
+}
