+++ src/components/ReciteBox/index.js
@@ -1,0 +1,3 @@
+import ReciteBox from './ReciteBox';
+
+export default ReciteBox;
